EE-225: Reorganize tests from `tests` crate.
This actually removes the crate completely and specific tests are moved
closer to the crate that defines things that are tested.
