+++ execution-engine/Cargo.toml
@@ -7,7 +7,6 @@ members = [
     "gens",
     "shared",
     "storage",
-    "tests",
     "vm",
     "wasm-prep",
 ]
