+++ execution-engine/common/Cargo.toml
@@ -15,3 +15,9 @@ failure = { version = "0.1.5", default-features = false, features = ["failure_de
 num = { version = "0.2.0", default-features = false }
 wee_alloc = "0.4.3"
 uint = { version = "0.6.1", default-features = false, features = [] }
+
+[dev-dependencies]
+common = { path = "../common", features = ["std"], package = "casperlabs-contract-ffi" }
+gens = { path = "../gens" }
+proptest = "0.9.2"
+shared = { path = "../shared" }
