+++ execution-engine/common/src/lib.rs
@@ -13,6 +13,9 @@ extern crate alloc;
 extern crate uint;
 extern crate failure;
 extern crate wee_alloc;
+#[cfg(test)]
+#[macro_use]
+extern crate proptest;
 
 #[cfg(not(feature = "std"))]
 #[global_allocator]
