--- execution-engine/tests/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "tests"
-version = "0.1.0"
-authors = ["Mateusz Górski <gorski.mateusz@protonmail.ch>"]
-
-[dependencies]
-common = { path = "../common", features = ["std"], package = "casperlabs-contract-ffi" }
-gens = { path = "../gens" }
-proptest = "0.9.2"
-shared = { path = "../shared" }
-
