--- execution-engine/tests/src/lib.rs
@@ -1,15 +1,0 @@
-#[cfg(test)]
-#[macro_use]
-extern crate proptest;
-#[cfg(test)]
-extern crate common;
-#[cfg(test)]
-extern crate gens;
-#[cfg(test)]
-extern crate shared;
-
-#[cfg(test)]
-mod bytesrepr;
-
-#[cfg(test)]
-mod partialord;
